src/requirements.txt
amqp==5.0.5
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3   
click-plugins==1.1.1      
click-repl==0.1.6
cryptography==3.3.1       
defusedxml==0.6.0
Django==3.1.3
django-allauth==0.44.0    
django-autoslug==1.9.8    
django-cas-ng==4.1.1      
django-crispy-forms==1.10.0
django-environ==0.4.5
django-mama-cas==2.4.0
django-model-utils==4.1.1
django-notifications-hq==1.6.0
django-simple-sso==0.14.1
django-tinymce==3.2.0
djangorestframework==3.12.2
idna==2.10
importlib-metadata==3.7.2
itsdangerous==1.1.0
jsonfield==3.1.0
kombu==5.0.2
lxml==4.6.2
oauthlib==3.1.0
Pillow==8.0.1
prompt-toolkit==3.0.16
pycparser==2.20
PyJWT==2.0.0
python-cas==1.5.0
python3-openid==3.2.0
pytz==2020.4
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
swapper==1.1.2.post1
typing-extensions==3.7.4.3
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
webservices==0.7
zipp==3.4.1